+++ src/main/java/dev/willyelton/crystal_tools/common/capability/Capabilities.java
@@ -6,8 +6,8 @@ import net.neoforged.neoforge.capabilities.ItemCapability;
 import static dev.willyelton.crystal_tools.CrystalTools.rl;
 
 public class Capabilities {
-    public static final ItemCapability<Levelable, RegistryAccess> ITEM_SKILL =
+    public static final ItemCapability<LevelableStack, RegistryAccess> ITEM_SKILL =
             ItemCapability.create(rl("item_skill"),
-                    Levelable.class,
+                    LevelableStack.class,
                     RegistryAccess.class);
 }