+++ src/main/java/dev/willyelton/crystal_tools/common/entity/CrystalTridentEntity.java
@@ -2,7 +2,7 @@ package dev.willyelton.crystal_tools.common.entity;
 
 import dev.willyelton.crystal_tools.Registration;
 import dev.willyelton.crystal_tools.common.capability.Capabilities;
-import dev.willyelton.crystal_tools.common.capability.Levelable;
+import dev.willyelton.crystal_tools.common.capability.LevelableStack;
 import dev.willyelton.crystal_tools.common.components.DataComponents;
 import dev.willyelton.crystal_tools.common.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.common.levelable.tool.CrystalTrident;
@@ -140,8 +140,8 @@ public class CrystalTridentEntity extends AbstractArrow {
             }
 
             if (damagingEntity instanceof Player player) {
-                Levelable levelable = tridentStack.getCapability(Capabilities.ITEM_SKILL, level().registryAccess());
-                if (levelable != null) {
+                LevelableStack levelable = tridentStack.getCapability(Capabilities.ITEM_SKILL, level().registryAccess());
+                if (levelable != null && levelable.allowDamageXp()) {
                     levelable.addExp(level(), damagingEntity.getOnPos(), player, damage);
                 }
             }