+++ src/main/java/dev/willyelton/crystal_tools/common/events/ProjectileHitEvent.java
@@ -2,7 +2,7 @@ package dev.willyelton.crystal_tools.common.events;
 
 import dev.willyelton.crystal_tools.CrystalTools;
 import dev.willyelton.crystal_tools.common.capability.Capabilities;
-import dev.willyelton.crystal_tools.common.capability.Levelable;
+import dev.willyelton.crystal_tools.common.capability.LevelableStack;
 import dev.willyelton.crystal_tools.utils.ToolUtils;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
@@ -27,14 +27,16 @@ public class ProjectileHitEvent {
                     event.getRayTraceResult() instanceof EntityHitResult hitResult) {
                 if (hitResult.getEntity() instanceof LivingEntity target && ToolUtils.isValidEntity(target)) {
                     ItemStack heldItem = player.getMainHandItem();
-                    Levelable levelable = heldItem.getCapability(Capabilities.ITEM_SKILL, player.level().registryAccess());
+                    LevelableStack levelable = heldItem.getCapability(Capabilities.ITEM_SKILL, player.level().registryAccess());
                     if (levelable != null) {
                         if (event.getProjectile() instanceof AbstractArrow arrow) {
                             float f = (float) arrow.getDeltaMovement().length();
                             // This is how they get damage, ignore crits for now
                             int damage = Mth.ceil(Mth.clamp((double) f * 2, 0.0D, 2.147483647E9D));
 
-                            levelable.addExp(player.level(), player.getOnPos(), player, (float) damage);
+                            if (levelable.allowDamageXp()) {
+                                levelable.addExp(player.level(), player.getOnPos(), player, (float) damage);
+                            }
                         }
                     }
                 }